+++ src/proof_system/widget/permutation/proverkey.rs
@@ -4,8 +4,8 @@
 //
 // Copyright (c) DUSK NETWORK. All rights reserved.
 
+use crate::composer::permutation::constants::{K1, K2, K3};
 use crate::fft::{EvaluationDomain, Evaluations, Polynomial};
-use crate::permutation::constants::{K1, K2, K3};
 use dusk_bls12_381::BlsScalar;
 
 #[cfg(feature = "rkyv-impl")]